WIP: Inflection point, scatter wrap.h towards plausible components
