+++ src/enclave/enclave.h
@@ -13,8 +13,9 @@
 #include "indexing/enclave_lfs_access.h"
 #include "indexing/historical_transaction_fetcher.h"
 #include "interface.h"
+#include "js/context.h"
+#include "js/ffi_plugins.h"
 #include "js/interpreter_cache.h"
-#include "js/wrap.h"
 #include "node/acme_challenge_frontend.h"
 #include "node/historical_queries.h"
 #include "node/network_state.h"
