+++ src/js/test/js.cpp
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache 2.0 License.
 #define DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN
-#include "js/wrap.cpp"
+#include "js/map_access_permissions.h"
 
 #include <doctest/doctest.h>
 
@@ -21,42 +21,42 @@ TEST_CASE("Check KV Map access")
     {
       INFO("Public internal tables are read-only");
       REQUIRE(
-        _check_kv_map_access(TxAccess::APP_RW, public_internal_table_name) ==
+        check_kv_map_access(TxAccess::APP_RW, public_internal_table_name) ==
         MapAccessPermissions::READ_ONLY);
     }
 
     {
       INFO("Private tables in internal namespace cannot even be read");
       REQUIRE(
-        _check_kv_map_access(TxAccess::APP_RW, private_internal_table_name) ==
+        check_kv_map_access(TxAccess::APP_RW, private_internal_table_name) ==
         MapAccessPermissions::ILLEGAL);
     }
 
     {
       INFO("Governance tables are read-only");
       REQUIRE(
-        _check_kv_map_access(TxAccess::APP_RW, public_gov_table_name) ==
+        check_kv_map_access(TxAccess::APP_RW, public_gov_table_name) ==
         MapAccessPermissions::READ_ONLY);
     }
 
     {
       INFO("Private tables in governance namespace cannot even be read");
       REQUIRE(
-        _check_kv_map_access(TxAccess::APP_RW, private_gov_table_name) ==
+        check_kv_map_access(TxAccess::APP_RW, private_gov_table_name) ==
         MapAccessPermissions::ILLEGAL);
     }
 
     {
       INFO("Public application tables are read-write");
       REQUIRE(
-        _check_kv_map_access(TxAccess::APP_RW, public_app_table_name) ==
+        check_kv_map_access(TxAccess::APP_RW, public_app_table_name) ==
         MapAccessPermissions::READ_WRITE);
 
       {
         INFO(
           "Unless the operation is read-only, in which case they're read-only");
         REQUIRE(
-          _check_kv_map_access(TxAccess::APP_RO, public_app_table_name) ==
+          check_kv_map_access(TxAccess::APP_RO, public_app_table_name) ==
           MapAccessPermissions::READ_ONLY);
       }
     }
@@ -64,14 +64,14 @@ TEST_CASE("Check KV Map access")
     {
       INFO("Private application tables are read-write");
       REQUIRE(
-        _check_kv_map_access(TxAccess::APP_RW, private_app_table_name) ==
+        check_kv_map_access(TxAccess::APP_RW, private_app_table_name) ==
         MapAccessPermissions::READ_WRITE);
 
       {
         INFO(
           "Unless the operation is read-only, in which case they're read-only");
         REQUIRE(
-          _check_kv_map_access(TxAccess::APP_RO, private_app_table_name) ==
+          check_kv_map_access(TxAccess::APP_RO, private_app_table_name) ==
           MapAccessPermissions::READ_ONLY);
       }
     }
@@ -82,42 +82,42 @@ TEST_CASE("Check KV Map access")
     {
       INFO("Public internal tables are read-only");
       REQUIRE(
-        _check_kv_map_access(TxAccess::GOV_RO, public_internal_table_name) ==
+        check_kv_map_access(TxAccess::GOV_RO, public_internal_table_name) ==
         MapAccessPermissions::READ_ONLY);
     }
 
     {
       INFO("Private tables in internal namespace cannot even be read");
       REQUIRE(
-        _check_kv_map_access(TxAccess::GOV_RO, private_internal_table_name) ==
+        check_kv_map_access(TxAccess::GOV_RO, private_internal_table_name) ==
         MapAccessPermissions::ILLEGAL);
     }
 
     {
       INFO("Governance tables are read-only");
       REQUIRE(
-        _check_kv_map_access(TxAccess::GOV_RO, public_gov_table_name) ==
+        check_kv_map_access(TxAccess::GOV_RO, public_gov_table_name) ==
         MapAccessPermissions::READ_ONLY);
     }
 
     {
       INFO("Private tables in governance namespace cannot even be read");
       REQUIRE(
-        _check_kv_map_access(TxAccess::GOV_RO, private_gov_table_name) ==
+        check_kv_map_access(TxAccess::GOV_RO, private_gov_table_name) ==
         MapAccessPermissions::ILLEGAL);
     }
 
     {
       INFO("Public application cannot even be read");
       REQUIRE(
-        _check_kv_map_access(TxAccess::GOV_RO, public_app_table_name) ==
+        check_kv_map_access(TxAccess::GOV_RO, public_app_table_name) ==
         MapAccessPermissions::ILLEGAL);
     }
 
     {
       INFO("Private application cannot even be read");
       REQUIRE(
-        _check_kv_map_access(TxAccess::GOV_RO, private_app_table_name) ==
+        check_kv_map_access(TxAccess::GOV_RO, private_app_table_name) ==
         MapAccessPermissions::ILLEGAL);
     }
   }
@@ -128,42 +128,42 @@ TEST_CASE("Check KV Map access")
     {
       INFO("Public internal tables are read-only");
       REQUIRE(
-        _check_kv_map_access(TxAccess::GOV_RW, public_internal_table_name) ==
+        check_kv_map_access(TxAccess::GOV_RW, public_internal_table_name) ==
         MapAccessPermissions::READ_ONLY);
     }
 
     {
       INFO("Private tables in internal namespace cannot even be read");
       REQUIRE(
-        _check_kv_map_access(TxAccess::GOV_RW, private_internal_table_name) ==
+        check_kv_map_access(TxAccess::GOV_RW, private_internal_table_name) ==
         MapAccessPermissions::ILLEGAL);
     }
 
     {
       INFO("Governance tables are read-write");
       REQUIRE(
-        _check_kv_map_access(TxAccess::GOV_RW, public_gov_table_name) ==
+        check_kv_map_access(TxAccess::GOV_RW, public_gov_table_name) ==
         MapAccessPermissions::READ_WRITE);
     }
 
     {
       INFO("Private tables in governance namespace cannot even be read");
       REQUIRE(
-        _check_kv_map_access(TxAccess::GOV_RW, private_gov_table_name) ==
+        check_kv_map_access(TxAccess::GOV_RW, private_gov_table_name) ==
         MapAccessPermissions::ILLEGAL);
     }
 
     {
       INFO("Public applications tables cannot even be read");
       REQUIRE(
-        _check_kv_map_access(TxAccess::GOV_RW, public_app_table_name) ==
+        check_kv_map_access(TxAccess::GOV_RW, public_app_table_name) ==
         MapAccessPermissions::ILLEGAL);
     }
 
     {
       INFO("Private applications tables cannot even be read");
       REQUIRE(
-        _check_kv_map_access(TxAccess::GOV_RW, private_app_table_name) ==
+        check_kv_map_access(TxAccess::GOV_RW, private_app_table_name) ==
         MapAccessPermissions::ILLEGAL);
     }
   }
