+++ src/node/node_state.h
@@ -25,7 +25,8 @@
 #include "history.h"
 #include "http/http_parser.h"
 #include "indexing/indexer.h"
-#include "js/wrap.h"
+#include "js/global_class_ids.h"
+#include "js/context.h"
 #include "network_state.h"
 #include "node/hooks.h"
 #include "node/http_node_client.h"
